LPS-30787 Regenerate

+++ portal-impl/src/com/liferay/portal/service/persistence/UserGroupPersistenceImpl.java
@@ -351,7 +351,9 @@ public class UserGroupPersistenceImpl extends BasePersistenceImpl<UserGroup>
 						userGroup.getPrimaryKeyObj());
 			}
 
-			session.delete(userGroup);
+			if (userGroup != null) {
+				session.delete(userGroup);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -360,7 +362,9 @@ public class UserGroupPersistenceImpl extends BasePersistenceImpl<UserGroup>
 			closeSession(session);
 		}
 
-		clearCache(userGroup);
+		if (userGroup != null) {
+			clearCache(userGroup);
+		}
 
 		return userGroup;
 	}
